EdkRepo: Fix sync -u when the initial combo is not present in the new manifest
Allow the sync command to update the local manifest when the
combos in the initial manifest are no longer present if a
combo with compatible repo_sources is available. If this case
is encountered change the current checked out combo to be the
default combination from the new manifest file.

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Puja Pandya <puja.pandya@intel.com>
Cc: Erik Bjorge <erik.c.bjorge@intel.com>
Cc: Bret Barkelew <Bret.Barkelew@microsoft.com>
Cc: Prince Agyeman <prince.agyeman@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>