EdkRepo: Update combo displayed in shell after checkout
With the current version of the command completion script,
after running "edkrepo checkout" the branch combination
listed in the shell prompt is not updated to reflect the
new branch combination. The reason for this is because
the branch combination isn't checked for changes unless
the present working directory was changed by the previous
command.

To fix this bug, the scripts have been updated to check
for changes to the branch combination when the last
command run by the user was "edkrepo" OR the present
working directory was changed by the previous command.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Ashley DeSimone <ashley.e.desimone@intel.com>
Cc: Puja Pandya <puja.pandya@intel.com>
Cc: Erik Bjorge <erik.c.bjorge@intel.com>
Cc: Prince Agyeman <prince.agyeman@intel.com>
Cc: Bret Barkelew <Bret.Barkelew@microsoft.com>
Cc: Philippe Mathieu-Daude <philmd@redhat.com>
Reviewed-by: Ashley DeSimone <ashley.e.desimone@intel.com>