EdkRepo: Address corner cases in find_project_in_all_indices
Update find_project_in_all_indices() to only return None values
for the source manifest repo and config file if the project is
truly not found in any manifest repository. In the case a path
to a manifest file is provided recursively call
find_in_all_indices() with the code name from the manifest at
the provided path to generate the source manifest repo and config
values.

Additionally update the call to find_project_in_all_indices() in
find_source_manifest_repo to use the value of man_repo instead of
None.

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Puja Pandya <puja.pandya@intel.com>
Cc: Erik Bjorge <erik.c.bjorge@intel.com>
Cc: Bret Barkelew <Bret.Barkelew@microsoft.com>
Cc: Prince Agyeman <prince.agyeman@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>