EdkRepo: Fixing inconsistent win32 HOME path mapping
In Python 3.8.x the method of mapping the users home path was changed
to look at USERPROFILE and then HOMEDRIVE\HOMEPATH.  This change removed
the check of HOME generated by git BASH based on HOMEDRIVE/HOMEPATH.
For users with a USERPROFILE environment variable that does not match
the HOMEDRIVE\HOMEPATH environment variables this can lead to
inconsistent decoding of files such as .gitconfig between BASH and
EdkRepo.

This change attempts to enforce a consistent decode order.
HOME, HOMEDRIVE\HOMEPATH and USERPROFILE.

Cc: Ashley E Desimone <ashley.e.desimone@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Puja Pandya <puja.pandya@intel.com>
Cc: Bret Barkelew <Bret.Barkelew@microsoft.com>
Cc: Prince Agyeman <prince.agyeman@intel.com>
Cc: Erik Bjorge <erik.c.bjorge@intel.com>
Signed-off-by: Erik Bjorge <erik.c.bjorge@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Ashley DeSimone <ashley.e.desimone@intel.com>