EdkRepo: Sync -u should support repository Local directory renames and/or URL changes
Updated comments to better describe the process of updating the
local manifest file.

Removed the exception that was raised when a URL change was
detected.

Updated the calculation of sources_to_move, sources_to_remove
and sources_to_clone so that the inital repository is moved
to an archival location and the updated respository is cloned.

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>