EdkRepo: Automatic remote prune in EdkRepo Sync
This resolves the "error: cannot lock ref 'refs/y': 'refs/x' exists;
cannot create 'refs/y'" message that is seen when a individual ref
is changed to a folder of refs. When this unusual even occurs, a
prune of existing refs needs to be done. It would be useful if
EdkRepo is able to detect this exceptional case and perform the prune
automatically.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>