EdkRepo: Handle broken source manifest repo links
If the project for a manifest file moves from one manifest repository to
another, then it is possible for the SourceManifestRepository under
GeneralConfig to refer to a manifest repository that no longer exists or
does not contain a project with the same codename.

This patch adds detection for this scenario. If it occurs, then a full
search of all known manifest repositories is conducted to attempt to
find the new location of the project's manifest. If the new manifest
directory is found, then the SourceManifestRepository is updated to
point to the new manifest repository.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>