Edkrepo: Bug Fixes
- Fixed the out of order parameters in clone_repos function.
- Changed names of methods from patchset_application_flow and
is_collision to patchset_branch_creation_flow and
is_branch_name_collision respectively.
- Removed the manifest_path variable in sync_command as it was
calculating the wrong path leading to a bug.
- Brought back the exception while creating a local branch for a branch
name that already exists.
- Removed magic strings from the code.
- While checking for collision, used 'git rev-parse' on the branch name
to get the Head instead of checking out to that branch.
- Implemented __eq__ method for _Patchset class.
- Removed the check_patchset_similarity method and used the equality
operator instead.
- When creating new repos at sync, changed the flow from creating and
then renaming to renaming and then creating which would prevent a
potential bug.
- If applying a patch fails, running a 'git am --abort' to cancel the
operation and then raise an exception.
- While applying cherry pick from a different remote, we now check to
see if the url in the manifest exists in the 'git remote -v' list.
- If a Cherry pick fails, we check for a merge conflict and execute
'git cherry-pick --abort' if there is a conflict.
- Removed ':' from the humble strings to make it consistent.

Signed-off-by: Harsh Vora harsh.vora@intel.com
Reviewed-by: Ashley DeSimone <ashley.e.desimone@intel.com>
Reviewed-by:  Kevin Sun <kevin.sun@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>