EdkRepo: Add selective cache updates to cache command
Add the 'selective' flag to the cache command; this flag is
only applicable when a manifest|pin file or project name is
passed via the project arg or when run from within a workspace.

Add a method to the cache command to determine which refs are
required to be fetched.

Extend the get_latest_sha() method in common repo functions to
work when no Repo object is available.

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>