Edkrepo: Bug Fixes and Improvements
- Implemented Patchset naming restriction to not use "main" or "master"
as names.
- Fixed the bug of edkrepo sync -u not working because of two file
pointers opening at the same time in create_repos and create_local_branch.
- Fixed a bug of named_tuples not being assigned in json.
- Implemented the functionality to take into account user created local
branches of the same name as patchset and to override and not create
branches using edkrepo if they want to use their own branches. If the
override flag is not specified, it throws an exception to tell the user
to manually resolve branch name conflicts.

Signed-off-by: Harsh Vora harsh.vora@intel.com
Reviewed-by: Ashley DeSimone <ashley.e.desimone@intel.com>
Reviewed-by:  Kevin Sun <kevin.sun@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>