Daisy-chaining PatchSet fails to create branch
Currently, two PatchSet can be chained together to create a dynamic branch.

If a third or more PatchSet is chained in a row, these PatchSet fail to create a dynamic branch when used, and non-gracefully fails when the PatchSet steps are attempted on the wrong branch or commit.

Locate the correct base PatchSet from the manifest before attempting to create the branch.
If a PatchSet circular dependency is found, raise an exception.
Do not attempt to create the branch if any exception besides KeyError is encountered.

Signed-off-by: Nathaniel Haller <nathaniel.d.haller@intel.com>
Reviewed-by: Kevin Sun <kevin.sun@intel.com>
Reviewed-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Nate Desimone <nathaniel.l.desimone@intel.com>