EdkRepo: Linux Installer Improvements
REF: #163

If bash-completion is available then it will now be used on Linux
systems. This resolves an issue on Raspberry Pi OS where installing
EdkRepo will cause the X11 server to crash during startup. The reason
this occurs is because the X11 startup scripts on Raspberry Pi blindly
runs all scripts in /etc/profile.d using /bin/sh

Adds support for user local installs on Linux and makes that the
recommended method.

Adds interactive installation mode to aid users in setting the
installation options correctly.

Fixes shell integration on Raspberry Pi OS.

Some of the shell initialization script additions made by the installer
are typically only needed on macOS. Moved those to only be added if the
current OS is macOS.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Kevin Sun <kevin.sun@intel.com>