EdkRepo: Unmask errors from __update_local_manifest() method.
Currently all errors from the __update_local_manifest() method
in the sync command are only visible when using the -v flag.
This prevents users from knowing why the manifest update was
not completed succesfully and that in many cases it can be re-run
and complete if the -o flag is used.

Fixes #166

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Kevin Sun <kevin.sun@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>