EdkRepo: List Repos Should Handle Combos That Do Not Reference a Branch
The list repos command will fail to process if any of the available
manifest files contain a combo which does not reference a branch.
Update the command to by-pass this case and to identify if a commit
or a tag is used in a combo. The output has been updated to show all
branches, commits or tags in a given repo that are used and which combo
they are used in.

Fixes: #167

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>