Selective Caching Does Not Function When Default Remote Name is Used
When the default remote name 'origin' is used for a cached repository
the cache's directory name will be the repository name not the remote
name; when this occurs the url_or_name parameter will match the value
of remote.name. This commit updates the conditional in update_cache() to
handle this case by checking for both possibilities.

Fixes #184

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Kevin Sun <kevin.sun@intel.com>