EdkRepo: Process Refs By Priority Order in List-Repos and Combo
In the List-Repos command a combo referencing a combination of tag,
sha, and branch would only have the branch listed. Update if / elif
block to ensure that all ref types are listed

In the combo commmand a combo containing a combination of tag,
sha and branch would only show the branch. Update the conditionals
such that the printed ref follows EdkRepo's priority order of ref
usage (SHA > Tag > Branch)

Fixes #182

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>
Reviewed-by: Kevin Sun <kevin.sun@intel.com>