Fixing default cache path restoration issue
When running cache commands with a custom path defined the default
path will be restored. This was due to the handling of cache commands
When the --path flag was not specified.

Updated the --path flag handling logic to do noting if not present. The
default path can be restored by passing 'default' as the PATH parameter.

Signed-off-by: Erik Bjorge <erik.c.bjorge@intel.com>