EdkRepo: Installer should perform a case-insensitive compare for PATH
REF: https://github.com/tianocore/edk2-edkrepo/issues/261

The EdkRepo installer does a check to make sure that the Windows directory
is contained within the PATH environment variable. However, retrieving the
current value of PATH is done via a string comparison with "Path". Since
environment variables are case-insensitive on Windows, the environment
variable could be named "Path", "path", or "PATH". If "Path" is not found,
then EdkRepo will clear and overwrite the current value of the PATH environment
variable at a system level. The fix for this issue is to perform a
case-insensitive comparison when searching for the PATH environment variable.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>