EdkRepo: Fix Combo Verbose Output When a Pin File is Checked Out
When a Pin file is checked out the value of the current cloned combo
is a custom string that indicates a pin file is in use; however, a
combo with this name is not present in the workspace's local manifest
file resulting in an error when accessing the combo's description.

In this case the error can be avoided by setting the value of the
description to None.

Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>