EdkRepo: Fix Display Of EdkRepo Combo In Git Bash
When running in the Visual Studio Code Terminal, the bash init script execution
sequence is different. This prevents the EdkRepo combo from being displayed
correctly when using Git Bash in the Visual Studio Code Terminal.

This change adds a check that ensures that the addition of the EdkRepo combo to
the prompt will always execute last, after all other init scripts are done.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>