Manifest Parser Support for Nested Repos
Add the 'nested_repo' entry to the RepoSource tuple definition
and update the _RepoSource class to populate this data based on
the defined path for the local_root.

Add the method 'get_parent_of_nested_repo' to the ManifestXML
class to determine if an appropriate parent repository exists.
The RepoSource tuple for the parent is returned otherwise a ValueError
will be raised if called with a non-nested repository or no parent repository
is found.

Add unit tests for the __RepoSource class's __init__ and tuple() methods.
Add unit tests for the get_parent_of_nested_repo() method.

resolves: #280
Signed-off-by: Ashley E Desimone <ashley.e.desimone@intel.com>