EdkRepo: Fix sync -u failure when manifest contains patchsets
When attempting to run a sync -u with a manifest file that contains one
or more patchsets, a AttributeError is currently generated by the
check_dirty_repos() function because the new manifest file does not have
a current_combo initialized yet. The fix is to use the default_combo if
the current_combo is None.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>