EdkRepo: Fix Intermittent Installer Hang
On some systems, the EdkRepo installer hangs forever without any log output.
This has been root caused to the Python interpreter detection logic. If the
user has the environment variable PYLAUNCHER_ALLOW_INSTALL or
PYLAUNCHER_ALWAYS_INSTALL set globally in their Windows registry, then
attempting to run "py --version" could cause the Python Launcher for Windows to
attempt to install Python from the Windows Store. The installation process
prompts the user interactively to ask if they wish to proceed. Since this is a
background process, the result is a hang.

The fix is to unset those environment variables so the Python Launcher for
Windows skips its automatic installation behavior.

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>