EdkRepo: Fix failure in checkout-pin
Checkout-pin currently fails with the following error:

The selected PIN file does not refer to the same project as the local manifest
file. Exiting without checkout out PIN data.

This is because it is comparing the entire RemoteRepo() tuple instead of just
the name and URL. The other fields of the RemoteRepo() tuple are currently
owner, review_type, and pr_strategy. These fields are inmaterial to the
compatibility of a pin file with a manifest file and should therefore be
excluded from the comparison.

Tracker-link: https://github.com/tianocore/edk2-edkrepo/issues/301

Signed-off-by: Nate DeSimone <nathaniel.l.desimone@intel.com>